requirements.txt
Flask
flask-cors
SQLAlchemy
psycopg2-binary
python-dotenv